refactor: simplify test and benchmark support
