+++ packages/example/package.json
@@ -3,7 +3,6 @@
     "private": true,
     "version": "0.0.0",
     "type": "module",
-    "scripts": {},
     "devDependencies": {
         "typescript": "catalog:",
         "vite": "catalog:",
