+++ packages/example/src/webfont/icons.ts
@@ -1,8 +1,8 @@
 export const iconBaseSelector = '.exIcon' as const;
 export const iconClassPrefix = 'icon-' as const;
 export const icons = [
-    'calendar',
     'add',
+    'calendar',
 ] as const;
 
 export type IconOption = typeof icons[number];
